meteo/requirements.txt
requests
cassandra-driver
kafka-python
streamlit
